Sincronizar

+++ SendDataDemo1/SendDataDemo-onDesktop/src/org/sunspotworld/demo/SendDataDemoHostApplication.java
@@ -66,7 +66,7 @@ public class SendDataDemoHostApplication {
                 rCon.receive(dg);
                 String addr = dg.getAddress();  // read sender's Id
                 long time = dg.readLong();      // read time of the reading
-                int val = dg.readInt();         // read the sensor value
+                double val = dg.readDouble();         // read the sensor value
                 System.out.println(fmt.format(new Date(time)) + "  from: " + addr + "   value = " + val);
             } catch (Exception e) {
                 System.err.println("Caught " + e +  " while reading sensor samples.");
